moved publication related functions to utils/publish.py
